Added Linear Interpolation
